Extract ProjectStore class into project_store.py
OOP refactor: moved all project logic (CRUD, scaffolding, templates,
clone) from store.py into ProjectStore class in project_store.py.

- store.py: 1058 → 723 lines (re-exports preserve backward compat)
- project_store.py: 413 lines with ProjectStore class
- DRY: _build_info and _persist eliminate 4 duplicate blocks
- Dispatch: source → creator method via dict lookup
- Dependency injection: connection_getter passed to constructor
- 91 tests pass, zero regressions
