+++ tests/test_projects.py
@@ -23,10 +23,11 @@ from signalpilot.gateway.gateway.models import (
     ProjectStorage,
     ProjectUpdate,
 )
-from signalpilot.gateway.gateway.store import (
-    PROJECTS_FILE,
+from signalpilot.gateway.gateway.store import PROJECTS_FILE
+from signalpilot.gateway.gateway.project_store import (
     _generate_dbt_project_yml,
     _generate_profiles_yml,
+    _store,
     create_project,
     delete_project,
     get_project,
@@ -126,7 +127,7 @@ class TestCreateProject:
 
     def test_create_new_project(self):
         conn = _make_connection("prod-pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             proj = create_project(ProjectCreate(
                 name="my-proj",
                 connection_name="prod-pg",
@@ -140,7 +141,7 @@ class TestCreateProject:
 
     def test_scaffolds_directory_structure(self):
         conn = _make_connection("prod-pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             proj = create_project(ProjectCreate(
                 name="scaffold-test",
                 connection_name="prod-pg",
@@ -155,13 +156,13 @@ class TestCreateProject:
 
     def test_duplicate_name_raises(self):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             create_project(ProjectCreate(name="dup", connection_name="pg"))
             with pytest.raises(ValueError, match="already exists"):
                 create_project(ProjectCreate(name="dup", connection_name="pg"))
 
     def test_missing_connection_raises(self):
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=None):
+        with patch.object(_store, "_get_connection", return_value=None):
             with pytest.raises(ValueError, match="not found"):
                 create_project(ProjectCreate(name="bad", connection_name="nope"))
 
@@ -169,7 +170,7 @@ class TestCreateProject:
         local_dir = Path(tempfile.mkdtemp())
         (local_dir / "dbt_project.yml").write_text("name: test")
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             proj = create_project(ProjectCreate(
                 name="linked-proj",
                 connection_name="pg",
@@ -186,7 +187,7 @@ class TestCreateProject:
         (local_dir / "dbt_project.yml").write_text("name: test")
         (local_dir / "models").mkdir()
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             proj = create_project(ProjectCreate(
                 name="copied-proj",
                 connection_name="pg",
@@ -202,7 +203,7 @@ class TestCreateProject:
     def test_create_github_project(self):
         conn = _make_connection("pg", "postgres")
         mock_run = MagicMock(returncode=0, stderr="", stdout="")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn), \
+        with patch.object(_store, "_get_connection", return_value=conn), \
              patch("subprocess.run", return_value=mock_run) as mock_sub:
             # Pre-create the dir with dbt_project.yml to simulate clone
             project_dir = Path(_test_dir) / "projects" / "gh-proj"
@@ -224,7 +225,7 @@ class TestCreateProject:
 
     def test_create_github_missing_url_raises(self):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             with pytest.raises(ValueError, match="git_url is required"):
                 create_project(ProjectCreate(
                     name="no-url",
@@ -235,7 +236,7 @@ class TestCreateProject:
     def test_create_github_clone_failure_raises(self):
         conn = _make_connection("pg", "postgres")
         mock_run = MagicMock(returncode=128, stderr="fatal: repo not found", stdout="")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn), \
+        with patch.object(_store, "_get_connection", return_value=conn), \
              patch("subprocess.run", return_value=mock_run):
             with pytest.raises(ValueError, match="git clone failed"):
                 create_project(ProjectCreate(
@@ -248,7 +249,7 @@ class TestCreateProject:
     def test_create_dbt_cloud_project(self):
         conn = _make_connection("pg", "postgres")
         mock_run = MagicMock(returncode=0, stderr="", stdout="")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn), \
+        with patch.object(_store, "_get_connection", return_value=conn), \
              patch("subprocess.run", return_value=mock_run):
             project_dir = Path(_test_dir) / "projects" / "cloud-proj"
             project_dir.mkdir(parents=True, exist_ok=True)
@@ -268,7 +269,7 @@ class TestCreateProject:
 
     def test_create_dbt_cloud_missing_url_raises(self):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             with pytest.raises(ValueError, match="git_url is required"):
                 create_project(ProjectCreate(
                     name="no-url-cloud",
@@ -278,7 +279,7 @@ class TestCreateProject:
 
     def test_create_local_missing_path_raises(self):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             with pytest.raises(ValueError, match="does not exist"):
                 create_project(ProjectCreate(
                     name="bad-local",
@@ -296,7 +297,7 @@ class TestListProjects:
 
     def test_returns_created_projects(self):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             create_project(ProjectCreate(name="a", connection_name="pg"))
             create_project(ProjectCreate(name="b", connection_name="pg"))
         projects = list_projects()
@@ -310,7 +311,7 @@ class TestGetProject:
 
     def test_get_existing(self):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             create_project(ProjectCreate(name="exists", connection_name="pg"))
         proj = get_project("exists")
         assert proj is not None
@@ -325,7 +326,7 @@ class TestUpdateProject:
 
     def test_update_description(self):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             create_project(ProjectCreate(name="upd", connection_name="pg"))
         result = update_project("upd", ProjectUpdate(description="new desc"))
         assert result is not None
@@ -333,7 +334,7 @@ class TestUpdateProject:
 
     def test_update_tags(self):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             create_project(ProjectCreate(name="tagged", connection_name="pg"))
         result = update_project("tagged", ProjectUpdate(tags=["prod", "analytics"]))
         assert result.tags == ["prod", "analytics"]
@@ -343,7 +344,7 @@ class TestUpdateProject:
 
     def test_update_preserves_other_fields(self):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             create_project(ProjectCreate(
                 name="preserve",
                 connection_name="pg",
@@ -360,7 +361,7 @@ class TestDeleteProject:
 
     def test_delete_managed_removes_directory(self):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             proj = create_project(ProjectCreate(name="to-delete", connection_name="pg"))
         project_dir = Path(proj.project_dir)
         assert project_dir.exists()
@@ -372,7 +373,7 @@ class TestDeleteProject:
         local_dir = Path(tempfile.mkdtemp())
         (local_dir / "dbt_project.yml").write_text("name: test")
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             create_project(ProjectCreate(
                 name="linked-del",
                 connection_name="pg",
@@ -410,7 +411,7 @@ class TestProjectsAPI:
 
     def test_create_and_get(self, client):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             resp = client.post("/api/projects", json={
                 "name": "api-proj",
                 "connection_name": "pg",
@@ -427,7 +428,7 @@ class TestProjectsAPI:
 
     def test_create_duplicate_returns_409(self, client):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             client.post("/api/projects", json={"name": "dup-api", "connection_name": "pg"})
             resp = client.post("/api/projects", json={"name": "dup-api", "connection_name": "pg"})
         assert resp.status_code == 409
@@ -438,7 +439,7 @@ class TestProjectsAPI:
 
     def test_update(self, client):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             client.post("/api/projects", json={"name": "upd-api", "connection_name": "pg"})
         resp = client.put("/api/projects/upd-api", json={"description": "updated"})
         assert resp.status_code == 200
@@ -446,7 +447,7 @@ class TestProjectsAPI:
 
     def test_delete(self, client):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             client.post("/api/projects", json={"name": "del-api", "connection_name": "pg"})
         resp = client.delete("/api/projects/del-api")
         assert resp.status_code == 204
@@ -459,7 +460,7 @@ class TestProjectsAPI:
 
     def test_scan(self, client):
         conn = _make_connection("pg", "postgres")
-        with patch("signalpilot.gateway.gateway.store.get_connection", return_value=conn):
+        with patch.object(_store, "_get_connection", return_value=conn):
             client.post("/api/projects", json={"name": "scan-api", "connection_name": "pg"})
         resp = client.post("/api/projects/scan-api/scan")
         assert resp.status_code == 200
